Turn set_facts into a include_vars
We're using a first_found lookup instead of a with_first_found loop
so that the 'paths' parameter can be used. This ensures that only
vars from the role are included, and not vars from a parent calling
role. This can happen when a parent role has a higher priority vars
file available for inclusion than the role it calls.